clean: user controller dir structure
