+++ config/routes.rb
@@ -1,9 +1,7 @@
 Rails.application.routes.draw do
   namespace :api do
     namespace :v1 do
-      namespace :users, path: "/" do
-        resources :users
-      end
+      resources :users
     end
   end
 end
