+++ test/controllers/api/v1/users_controller_test.rb
@@ -1,6 +1,6 @@
 require "test_helper"
 
-class Api::V1::Users::UsersControllerTest < ActionDispatch::IntegrationTest
+class Api::V1::UsersControllerTest < ActionDispatch::IntegrationTest
   def setup
     @valid_user = { name: "test_user",
                     email: "test@testmail.com",
@@ -11,21 +11,21 @@ class Api::V1::Users::UsersControllerTest < ActionDispatch::IntegrationTest
 
   test "test creation of user " do
     assert_difference "User.count", 1 do
-      post api_v1_users_users_path, params: { user: { access_token: @token } }
+      post api_v1_users_path, params: { user: { access_token: @token } }
     end
     assert_response :ok
     assert_equal @valid_user[:name], JSON.parse(response.body)["user"]["name"]
   end
 
   test "test failer with duplicated user" do
-    post api_v1_users_users_path, params: { user: { access_token: @token } }
+    post api_v1_users_path, params: { user: { access_token: @token } }
 
     assert_raises ActiveRecord::RecordNotUnique do
-      post api_v1_users_users_path, params: { user: { access_token: @token } }
+      post api_v1_users_path, params: { user: { access_token: @token } }
     end
   end
   test "test failure with missing access_token" do
-    post api_v1_users_users_path, params: { user: { access_token: nil } }
+    post api_v1_users_path, params: { user: { access_token: nil } }
     assert_response :bad_request
     assert_equal "access token is missing", JSON.parse(response.body)["error"]
   end
